Consolidate client/server impls
