+++ minidumper/src/ipc/server.rs
@@ -1,0 +1,295 @@
+use super::{Connection, Header, Listener, SocketName};
+use crate::Error;
+
+/// Server side of the connection, which runs in the watchdog process that is
+/// meant to monitor the process where the [`super::Client`] resides
+pub struct Server {
+    listener: Listener,
+}
+
+struct ClientConn {
+    /// The actual socket connection we established with accept
+    socket: Connection,
+    /// The key we associated with the socket
+    key: usize,
+}
+
+impl ClientConn {
+    #[inline]
+    fn new(socket: Connection, key: usize) -> Self {
+        Self { socket, key }
+    }
+
+    fn recv(&mut self, handler: &dyn crate::ServerHandler) -> Option<(u32, Vec<u8>)> {
+        use std::io::IoSliceMut;
+
+        let mut hdr_buf = [0u8; std::mem::size_of::<Header>()];
+        cfg_if::cfg_if! {
+            if #[cfg(any(target_os = "linux", target_os = "android"))] {
+                let (len, _trunc) = self.socket.peek(&mut hdr_buf).ok()?;
+            } else {
+                let len = self.socket.peek(&mut hdr_buf).ok()?;
+            }
+        }
+
+        if len == 0 {
+            return None;
+        }
+
+        let header = Header::from_bytes(&hdr_buf)?;
+        let mut buffer = handler.message_alloc();
+
+        buffer.resize(header.size as usize, 0);
+
+        self.socket
+            .recv_vectored(&mut [IoSliceMut::new(&mut hdr_buf), IoSliceMut::new(&mut buffer)])
+            .ok()?;
+
+        Some((header.kind, buffer))
+    }
+}
+
+impl Server {
+    /// Creates a new server with the given name.
+    ///
+    /// Note that in the case of a path socket name, this method always attempts
+    /// to delete the specified path if it exists as both Windows and Macos have
+    /// issues around cleaning up these files if the process the server runs in
+    /// aborts abnormally.
+    ///
+    /// # Errors
+    ///
+    /// The provided socket name is invalid, or the listener socket was unable
+    /// to be bound to the specified socket name.
+    pub fn with_name<'scope>(name: impl Into<SocketName<'scope>>) -> Result<Self, Error> {
+        let sn = name.into();
+
+        #[allow(irrefutable_let_patterns)]
+        if let SocketName::Path(path) = &sn {
+            if path.exists() {
+                // We ignore the result, if we couldn't delete the file, it's
+                // most likely that the following bind will fail
+                let _res = std::fs::remove_file(path);
+            }
+        }
+
+        cfg_if::cfg_if! {
+            if #[cfg(any(target_os = "linux", target_os = "android"))] {
+                let socket_addr = match sn {
+                    SocketName::Path(path) => {
+                        uds::UnixSocketAddr::from_path(path).map_err(|_err| Error::InvalidName)?
+                    }
+                    SocketName::Abstract(name) => {
+                        uds::UnixSocketAddr::from_abstract(name).map_err(|_err| Error::InvalidName)?
+                    }
+                };
+
+                let listener = Listener::bind_unix_addr(&socket_addr)?;
+            } else {
+                let SocketName::Path(path) = sn;
+                let listener = Listener::bind(path)?;
+                listener.set_nonblocking(true)?;
+            }
+        }
+
+        Ok(Self { listener })
+    }
+
+    /// Runs the server loop, accepting client connections and receiving IPC
+    /// messages.
+    ///
+    /// # Errors
+    ///
+    /// This method uses basic I/O event notification via [`polling`] which
+    /// can fail for a number of different reasons
+    pub fn run(
+        &self,
+        handler: Box<dyn crate::ServerHandler>,
+        shutdown: &std::sync::atomic::AtomicBool,
+    ) -> Result<(), Error> {
+        use polling::{Event, Poller};
+
+        let poll = Poller::new()?;
+        let mut events = Vec::new();
+
+        poll.add(&self.listener, Event::readable(0))?;
+
+        let mut clients = Vec::new();
+        let mut id = 1;
+
+        loop {
+            if shutdown.load(std::sync::atomic::Ordering::Relaxed) {
+                return Ok(());
+            }
+
+            events.clear();
+            poll.wait(&mut events, Some(std::time::Duration::from_millis(10)))?;
+
+            for event in events.iter() {
+                if event.key == 0 {
+                    match self.listener.accept_unix_addr() {
+                        Ok((accepted, _addr)) => {
+                            let key = id;
+                            id += 1;
+
+                            poll.add(&accepted, Event::readable(key))?;
+
+                            log::debug!("accepted connection {}", key);
+                            clients.push(ClientConn::new(accepted, key));
+                        }
+                        Err(err) => {
+                            log::error!("failed to accept socket connection: {}", err);
+                        }
+                    }
+
+                    // We need to reregister insterest every time
+                    poll.modify(&self.listener, Event::readable(0))?;
+                } else if let Some(pos) = clients.iter().position(|cc| cc.key == event.key) {
+                    let deregister = match clients[pos].recv(handler.as_ref()) {
+                        Some((0, crash_context)) => {
+                            let cc = clients.swap_remove(pos);
+
+                            let exit = match Self::handle_crash_request(
+                                &cc.socket,
+                                &crash_context,
+                                handler.as_ref(),
+                            ) {
+                                Err(err) => {
+                                    log::error!("failed to capture minidump: {}", err);
+                                    false
+                                }
+                                Ok(exit) => {
+                                    log::info!("captured minidump");
+                                    exit
+                                }
+                            };
+
+                            if let Err(e) = cc.socket.send(&[1]) {
+                                log::error!("failed to send ack: {}", e);
+                            }
+
+                            if exit {
+                                return Ok(());
+                            }
+
+                            Some(cc.socket)
+                        }
+                        Some((kind, buffer)) => {
+                            handler.on_message(
+                                kind - 1, /* give the user back the original code they specified */
+                                buffer,
+                            );
+
+                            // We only send acks for crash dump requests
+                            // if let Err(e) = clients[pos].socket.send(&[1]) {
+                            //     log::error!("failed to send ack: {}", e);
+                            // }
+
+                            None
+                        }
+                        None => {
+                            log::debug!("client closed socket {}", pos);
+                            let cc = clients.swap_remove(pos);
+                            Some(cc.socket)
+                        }
+                    };
+
+                    if let Some(socket) = deregister {
+                        if let Err(e) = poll.delete(&socket) {
+                            log::error!("failed to deregister socket: {}", e);
+                        }
+                    } else {
+                        poll.modify(&clients[pos].socket, Event::readable(clients[pos].key))?;
+                    }
+                }
+            }
+        }
+    }
+
+    fn handle_crash_request(
+        _socket: &Connection,
+        buffer: &[u8],
+        handler: &dyn crate::ServerHandler,
+    ) -> Result<bool, Error> {
+        cfg_if::cfg_if! {
+            if #[cfg(any(target_os = "linux", target_os = "android"))] {
+                let peer_creds = _socket.initial_peer_credentials()?;
+
+                let pid = peer_creds.pid().ok_or(Error::UnknownClientPid)?;
+
+                let cc = crash_context::CrashContext::from_bytes(buffer).ok_or_else(|| {
+                    Error::from(std::io::Error::new(
+                        std::io::ErrorKind::InvalidData,
+                        "client sent an incorrectly sized buffer",
+                    ))
+                })?;
+
+                let mut writer =
+                    minidump_writer::minidump_writer::MinidumpWriter::new(pid.get() as i32, cc.tid);
+                writer.set_crash_context(minidump_writer::crash_context::CrashContext { inner: cc });
+            } else if #[cfg(target_os = "windows")] {
+                use scroll::Pread;
+                let dump_request: super::DumpRequest = buffer.pread(0)?;
+
+                // MiniDumpWriteDump primarily uses `EXCEPTION_POINTERS` for its crash
+                // context information, but inside that is an `EXCEPTION_RECORD`, which
+                // is an internally linked list, so rather than recurse and allocate until
+                // the end of that linked list, we just retrieve the actual pointer from
+                // the client process, and inform the dump writer that they are pointers
+                // to a different process, as MiniDumpWriteDump will internally read
+                // the processes memory as needed
+                let exception_pointers = dump_request.exception_pointers as *const std::ffi::c_void;
+
+                let cc = crash_context::CrashContext {
+                    exception_pointers,
+                    thread_id: dump_request.thread_id,
+                    exception_code: dump_request.exception_code,
+                };
+
+                let writer = minidump_writer::minidump_writer::MinidumpWriter::external_process(
+                    cc,
+                    dump_request.process_id,
+                )?;
+            } else if #[cfg(target_os = "macos")] {
+                use scroll::Pread;
+                let dump_request: super::DumpRequest = buffer.pread(0)?;
+
+                let exception = if dump_request.has_exception != 0 {
+                    Some(crash_context::ExceptionInfo {
+                        kind: dump_request.kind,
+                        code: dump_request.code,
+                        subcode: (dump_request.has_subcode != 0).then(|| dump_request.subcode),
+                    })
+                } else {
+                    None
+                };
+
+                let cc = crash_context::CrashContext {
+                    task: dump_request.task,
+                    thread: dump_request.thread,
+                    handler_thread: dump_request.handler_thread,
+                    exception,
+                };
+
+                let mut writer = minidump_writer::minidump_writer::MinidumpWriter::new(cc);
+            }
+        }
+
+        let (mut minidump_file, minidump_path) = handler.create_minidump_file()?;
+        let result = writer.dump(&mut minidump_file);
+
+        // Notify the user handler about the minidump, even if we failed to write it
+        Ok(handler.on_minidump_created(
+            result
+                .map(|_contents| crate::MinidumpBinary {
+                    file: minidump_file,
+                    path: minidump_path,
+                    #[cfg(target_os = "windows")]
+                    contents: Vec::new(),
+                    #[cfg(not(target_os = "windows"))]
+                    contents: _contents,
+                })
+                .map_err(crate::Error::from),
+        ))
+    }
+}
