--- minidumper/src/linux.rs
@@ -1,5 +1,0 @@
-mod client;
-mod server;
-
-pub use client::Client;
-pub use server::Server;
